refactor: Co-locate `page_html()` with the other `page_*()` functions in `_page.py`
